Search for older tweets with exact text
Detector now knows how to search for the tweet's text to see if
there are matches. If there are, it will keep searching with
max_id set to the lowest ID of the previous results until it finds
the oldets tweet.

There are some problems that still need to be fixed, including:

* The text must be EXACT.
* Line breaks aren't handled well.
* Quotes in the original tweet cause some voodoo to happen.
* Simple tweets (i.e. "I'm dying :joy::joy:") will have false
  positives.

Still, this is a good start.